Revert "Development to FromThePage"
