+++ app/assets/stylesheets/sections/collection.scss
@@ -403,64 +403,6 @@
   }
 }
 
-
-// Collection export
-
-.bulk-export {
-  padding: 0;
-  margin: 1.5em 0;
-  list-style: none;
-
-  li {
-    margin: 0;
-    display: flex;
-    padding: 15px 0;
-    border-style: solid;
-    border-width: 1px 0 0 0;
-    border-color: rgba(#000, 0.15);
-
-    &:last-child {
-      border-width: 1px 0;
-    }
-  }
-
-  &_format {
-    flex: 1 1 auto;
-    padding-right: 15px;
-  
-    h5 {
-      display: flex;
-      align-items: center;
-      margin: 0 0 0.25em 0;
-  
-      input {
-        width: 19px;
-        height: 19px;
-        margin-right: 5px;
-      }
-    }
-  }
-
-  &_options {
-    display: flex;
-    flex: 0 0 180px;
-    padding-left: 15px;
-    flex-direction: column;
-    justify-content: center;
-    border-left: 1px solid rgba(#000, 0.15);
-
-    label {
-      margin: 3px 0;
-      display: flex;
-      align-items: center;
-
-      input {
-        margin-right: 3px;
-      }
-    }
-  }
-}
-
 //Edit transcription fields
 .button.fields {
   margin: $gapSize/2 $gapSize/2 0 0;
