--- app/helpers/bulk_exports_helper.rb
@@ -1,2 +1,0 @@
-module BulkExportsHelper
-end
