+++ app/mailers/user_mailer.rb
@@ -1,5 +1,4 @@
 class UserMailer < ActionMailer::Base
-  include Rails.application.routes.url_helpers
   default from: SENDING_EMAIL_ADDRESS
   layout "mailer"
 
@@ -14,12 +13,6 @@ class UserMailer < ActionMailer::Base
     @document_upload = document_upload
 
     mail to: @document_upload.user.email, subject: "Your upload is ready"
-  end
-
-  def bulk_export_finished(bulk_export)
-    @bulk_export = bulk_export
-
-    mail to: @bulk_export.user.email, subject: "Your export is ready"
   end
 
   def new_owner(user, text)
