+++ app/models/collection.rb
@@ -14,7 +14,6 @@ class Collection < ApplicationRecord
   has_many :deeds, -> { order 'deeds.created_at DESC' }, :dependent => :destroy
   has_one :sc_collection, :dependent => :destroy
   has_many :transcription_fields, :dependent => :destroy
-  has_many :bulk_exports, :dependent => :destroy
 
   belongs_to :next_untranscribed_page, foreign_key: 'next_untranscribed_page_id', class_name: "Page", optional: true
   has_many :pages, through: :works
