+++ app/models/user.rb
@@ -300,14 +300,4 @@ class User < ApplicationRecord
 
     self.save
   end
-
-
-  # Generate a unique API key
-  def self.generate_api_key
-    loop do
-      token = SecureRandom.base64.tr('+/=', 'Qrt')
-      break token unless User.exists?(api_key: token)
-    end
-  end
-
 end
