+++ app/views/admin/_header.html.slim
@@ -8,6 +8,5 @@ h1 =t('.administration')
   =active_link_to t('.users'), admin_user_list_path
   =active_link_to t('.abuse'), admin_flag_list_path
   =active_link_to t('.uploads'), admin_uploads_path
-  =active_link_to t('.exports'), bulk_export_index_path
   =active_link_to t('.logfile'), admin_tail_logfile_path
   =active_link_to t('.settings'), admin_settings_path
