--- app/views/bulk_export/new.html.slim
@@ -1,119 +1,0 @@
-=render({ :partial => '/shared/collection_tabs', :locals => { :selected => 6, :collection_id => @collection }})
-
-.columns
-  article.maincol
-    h2.nomargin Export All Works
-    =form_with(url: bulk_export_create_path, model: @bulk_export, local: true) do |f|
-      =hidden_field_tag :collection_id, @collection.slug
-      =validation_summary @bulk_export.errors
-      ul.bulk-export
-        li
-          .bulk-export_format
-            h5 HTML
-            p.nomargin This may be useful for preservation in other systems or as a starting point for display on another website.
-          .bulk-export_options
-            =f.label :html_page
-              =f.check_box :html_page, value: @bulk_export.html_page
-              span One file per page
-            =f.label :html_work
-              =f.check_box :html_work, value: @bulk_export.html_work
-              span One file per work
-        li
-          .bulk-export_format
-            h5 Verbatim Plaintext
-            p.nomargin This plaintext file will represent line breaks with single newline, paragraph breaks with a double newline, and page breaks with a triple newline. It will contain the verbatim text, with all formatting, emendation, and subject linking stripped out.
-          .bulk-export_options
-            =f.label :plaintext_verbatim_page
-              =f.check_box :plaintext_verbatim_page, value: @bulk_export.plaintext_verbatim_page
-              span One file per page
-            =f.label :plaintext_verbatim_work
-              =f.check_box :plaintext_verbatim_work, value: @bulk_export.plaintext_verbatim_work
-              span One file per work
-        li
-          .bulk-export_format
-            h5 Expanded Plaintext
-            p.nomargin Like the verbatim plaintext, this plaintext file will represent line breaks with single newline, paragraph breaks with a double newline, and page breaks with a triple newline. It differs from the verbatim text, in that normalization will be applied to all subjects mentioned, so that while the verbatim text may read “I greeted Mr. Jones and his wife this morning.”, the emended plaintext will read “I greeted James Jones and Elizabeth Smith Jones this morning”. This artificial text is useful for programmatic analysis, but is not meant to be read by humans.
-          .bulk-export_options
-            =f.label :plaintext_emended_page
-              =f.check_box :plaintext_emended_page, value: @bulk_export.plaintext_emended_page
-              span One file per page
-            =f.label :plaintext_emended_work
-              =f.check_box :plaintext_emended_work, value: @bulk_export.plaintext_emended_work
-              span One file per work
-        li
-          .bulk-export_format
-            h5 Search-optimized Plaintext
-            p.nomargin A plaintext version of the work optimized for full-text search. This version contains a verbatim plaintext transcript of each page (as described above), except that words broken by hyphenated newlines are joined together, and a list of the canonical names mentioned within each page is appended to the end of the page.
-          .bulk-export_options
-            =f.label :plaintext_searchable_page
-              =f.check_box :plaintext_searchable_page, value: @bulk_export.plaintext_searchable_page
-              span One file per page
-            =f.label :plaintext_searchable_work
-              =f.check_box :plaintext_searchable_work, value: @bulk_export.plaintext_searchable_work
-              span One file per work
-        li
-          .bulk-export_format
-            h5 TEI XML
-            p.nomargin This may be useful to editors who plan to do further mark-up within TEI-XML editors like oXygen.
-          .bulk-export_options
-            =f.label :tei_work
-              =f.check_box :tei_work, value: @bulk_export.tei_work
-              span One file per work
-        li
-          .bulk-export_format
-            h5 Table/Field CSV
-            p.nomargin Exports a spreadsheet with field-based or tabular data.
-          .bulk-export_options
-            =f.label :table_csv_work
-              =f.check_box :table_csv_work, value: @bulk_export.table_csv_work
-              span One file per work
-            =f.label :table_csv_collection
-              =f.check_box :table_csv_collection, value: @bulk_export.table_csv_collection
-              span One file per collection
-        li
-          .bulk-export_format
-            h5 Subject CSV
-            p.nomargin A spreadsheet listing each place a subject is mentioned within a page within the collection
-          .bulk-export_options
-            =f.label :subject_csv_collection
-              =f.check_box :subject_csv_collection, value: @bulk_export.subject_csv_collection
-              span One file per collection
-
-      .toolbar
-        .toolbar_group.aright =f.button 'Start Export', class: 'big', 'data-start-export': ''
-
-  aside.sidecol
-    h4 Previous Exports
-    table.admin-grid.datagrid.striped
-      thead
-        tr
-          th Time
-          th Status
-
-      tbody
-        -@collection.bulk_exports.where.not(:status => BulkExport::Status::CLEANED).order(:created_at => :desc).each do |bulk_export|
-          tr
-            td.nowrap
-              =bulk_export.created_at.localtime.strftime("%m/%d/%Y at %I:%M %p")
-            td.nowrap
-              small.label(class="upload-status-#{bulk_export.status}") =bulk_export.status.titleize
-              -if bulk_export.status == BulkExport::Status::FINISHED
-                br
-                =link_to(t('.download'), bulk_export_download_path(:bulk_export_id => bulk_export.id), class: 'btnExport')
-
--content_for :javascript
-  javascript:
-    $(function() {
-      // Remove 'page_busy' spinner once download finished
-      // Cookie should be set to 'true' in the controller method
-      $('.btnExport').on('click', function() {
-        Cookies.remove('download_finished', {path: '/'});
-        var downloadCheckTimer = window.setInterval(function() {
-          var cookie = Cookies.get('download_finished');
-          if(cookie === 'true') {
-            $('html').removeClass('page-busy');
-            window.clearInterval(downloadCheckTimer);
-          }
-        }, 1000);
-      });
-    });
