+++ app/views/transcribe/display_page.html.slim
@@ -467,35 +467,19 @@ h2.legend =t('notes_and_questions')
           {
             hotkey: 'Ctrl-E',
             class: 'italic',
-            label: 'abbr',
+            label: 'expan',
             callback: function (cm) {
               var selection = cm.getSelection();
-              cm.replaceSelection('<abbr expan="">' + selection + '</abbr>');
+              cm.replaceSelection('<expan orig="">' + selection + '</expan>');
               if (!selection) {
                 var cursorPos = cm.getCursor();
-                cm.setCursor(cursorPos.line, cursorPos.ch - 7);
+                cm.setCursor(cursorPos.line, cursorPos.ch - 8);
               } else {
                 var cursorPos = cm.getCursor();
-                cm.setCursor(cursorPos.line, cursorPos.ch - 7 - selection.length - 2);
+                cm.setCursor(cursorPos.line, cursorPos.ch - 8 - selection.length - 2);
               }
             }
           },
-          // {
-          //   hotkey: 'Ctrl-E',
-          //   class: 'italic',
-          //   label: 'expan',
-          //   callback: function (cm) {
-          //     var selection = cm.getSelection();
-          //     cm.replaceSelection('<expan orig="">' + selection + '</expan>');
-          //     if (!selection) {
-          //       var cursorPos = cm.getCursor();
-          //       cm.setCursor(cursorPos.line, cursorPos.ch - 8);
-          //     } else { 
-          //       var cursorPos = cm.getCursor();
-          //       cm.setCursor(cursorPos.line, cursorPos.ch - 8 - selection.length - 2);
-          //     }
-          //   }
-          // },
           {
             hotkey: 'Ctrl-D',
             class: 'strikethrough',
