+++ app/views/user/profile.html.slim
@@ -1,6 +1,7 @@
 -user_name = @user.display_name.presence || @user.login
 -content_for :page_title, user_name
 -content_for :meta_description, "#{user_name}'s profile at FromThePage. #{@user.about}"
+
 -if @user.owner?
   =render partial: 'owner_profile', locals: { user: @user, carousel_collections: @carousel_collections, collections_and_document_sets: @collections_and_document_sets }
 -else
