--- app/views/user_mailer/bulk_export_finished.html.erb
@@ -1,5 +1,0 @@
-<h2 style="color:#453939; margin:0; font-weight:normal;">Your export is ready!</h2>
-<p>
-  Your export has been processed and is ready at
-  <%= link_to('this link.', bulk_export_download_url(@bulk_export)) %>.
-</p>
