--- app/views/user_mailer/bulk_export_finished.text.erb
@@ -1,4 +1,0 @@
-Your export is ready!
-
-Your export has been processed and is ready at this URL:
-<%= bulk_export_download_url(@bulk_export) %>
