+++ config/environments/development.rb
@@ -36,7 +36,7 @@ Rails.application.configure do
 
   config.action_mailer.perform_caching = false
 
-  Rails.application.routes.default_url_options = config.action_mailer.default_url_options =  { host: 'localhost:3000' }
+  config.action_mailer.default_url_options =  { host: 'localhost:3000' }
 
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.smtp_settings = { :address => "localhost", :port => 1025 }
