--- db/migrate/20210121203828_add_api_key_to_user.rb
@@ -1,5 +1,0 @@
-class AddApiKeyToUser < ActiveRecord::Migration[6.0]
-  def change
-    add_column :users, :api_key, :string
-  end
-end
